Added code compression, full LCD printing, & LEDs.
Compression added by Project > Options for target ' '> C/C++ > Optimization (0 = none)
Potentiometer currently affecting luminosity & temp when 0 somehow. Need to investigate further.
Need to discuss whether threads should be split further.

+++ lab_103/3_rtos_application/inc/main.h
@@ -21,7 +21,7 @@
 // mail data structure
 typedef struct 
 {
-  float     potVal;
+  uint16_t   potVal;
   uint8_t    ldrVal;
   float		  tempVal;
 }
